requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-socketio==5.8.0
python-engineio==4.8.0
eventlet==0.33.3
bcrypt==4.1.2
python-dotenv==1.0.1
